Create raspberry pi setup script
This will ensure the device has chromium installed, install a script under LDXE-pi autostart directory. Which will disable energy saving features and screenshots using xset, hide the mouse cursor with unclutter and start chromium fullscreen.

It seems the raspberry pi image defaults to underscan, so this will also ask the user if they see a black border, if so we will update the boot/config.txt file appropriately.

This allows a customer to setup a raspberry pi to act like a chromebit in kiosk - without the need to pay for a licence to do so.